requirements.txt
langchain~=0.3.26
google-genai
python-dotenv~=1.1.1
h11~=0.16.0
pip~=25.1.1
attrs~=25.3.0
protobuf~=5.29.5
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.73.1
requests~=2.32.4
typing_extensions~=4.14.1
exceptiongroup~=1.3.0
sniffio~=1.3.1
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
proto-plus~=1.26.1
aiohttp~=3.12.13
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.6.15
urllib3~=2.5.0
packaging~=24.2
google-api-python-client~=2.175.0
pydantic~=2.11.7
websockets~=15.0.1
anyio~=4.9.0
tenacity~=8.5.0
multidict~=6.6.3
grpcio-status~=1.71.2
httplib2~=0.22.0
tqdm~=4.67.1
pyparsing~=3.2.3
pydantic_core~=2.33.2
annotated-types~=0.7.0
charset-normalizer~=3.4.2
async-timeout~=4.0.3
uritemplate~=4.2.0
google-auth-httplib2~=0.2.0
serpapi~=0.1.5
langchain-community~=0.3.27
langchain-google-genai~=2.1.6